Ensure qb newb modal only appears once. Move currentUser reducer to metabase/redux
