+++ frontend/src/metabase/redux/user.js
@@ -1,9 +1,6 @@
 import { createAction } from "redux-actions";
 import { handleActions } from 'redux-actions';
 
-
-export const setUser = createAction("SET_USER");
-
 export const refreshCurrentUser = createAction("REFRESH_CURRENT_USER", async function getCurrentUser() {
     try {
         let response = await fetch("/api/user/current", { credentials: 'same-origin' });
@@ -16,8 +13,8 @@ export const refreshCurrentUser = createAction("REFRESH_CURRENT_USER", async fun
     return null;
 })
 
-export const currentUser = handleActions({
-    ["SET_USER"]: { next: (state, { payload }) => payload },
-    ["REFRESH_CURRENT_USER"]: { next: (state, { payload }) => payload },
-    ["AUTH_LOGOUT"]: { next: (state, { payload }) => null }
+export default handleActions({
+    ["REFRESH_CURRENT_USER"]:   { next: (state, { payload }) => payload },
+    ["AUTH_LOGOUT"]:            { next: (state) => null },
+    ["CLOSE_QB_NEWB_MODAL"]:    { next: (state) => state && { ...state, is_qbnewb: false } },
 }, null);
